create method for creating order

+++ src/main/java/org/example/logisticapplication/domain/RoutePoint/RoutePointDto.java
@@ -9,7 +9,7 @@ public record RoutePointDto(
         @Null
         Long id,
         String city,
-        Integer cargoId,
+        Long cargoId,
         String operationType
 ) {
 }
